+++ src/main/java/com/hunteryavitz/blockchainapi/services/BlockchainService.java
@@ -3,6 +3,7 @@ package com.hunteryavitz.blockchainapi.services;
 import com.hunteryavitz.blockchainapi.entities.Block;
 import com.hunteryavitz.blockchainapi.entities.Transaction;
 import com.hunteryavitz.blockchainapi.utils.Utils;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.security.NoSuchAlgorithmException;
@@ -15,6 +16,12 @@ import java.util.Arrays;
 public class BlockchainService {
 
     /**
+     * The health metric service.
+     */
+    @Autowired
+    HealthMetricService healthMetricService;
+
+    /**
      * The blockchain is an array of blocks.
      */
     private static Block[] blockchain;
@@ -25,14 +32,28 @@ public class BlockchainService {
     private static Boolean[] liveness;
 
     /**
+     * The blockCount is the number of blocks for the purposes of measuring production.
+     */
+    private static Integer[] contaminationSpectrum;
+
+    /**
      * The constructor for the BlockchainService class.
      */
     public void createInitialBlockchain() {
+
+        if (healthMetricService == null) {
+            healthMetricService = new HealthMetricService();
+        }
 
         liveness = new Boolean[100];
 
         blockchain = new Block[100];
         Block genesisBlock;
+
+        contaminationSpectrum = new Integer[5];
+        Arrays.fill(contaminationSpectrum, 0);
+
+        healthMetricService.resetBlockCount();
 
         try {
             long timestamp = System.currentTimeMillis();
@@ -43,6 +64,7 @@ public class BlockchainService {
         }
 
         blockchain[0] = genesisBlock;
+        healthMetricService.incrementBlockCount();
 
     }
 
@@ -70,6 +92,8 @@ public class BlockchainService {
                     System.currentTimeMillis(), "Block " + nextBlockIndex,
                     Utils.calculateHash(nextBlockIndex, previousBlock.getHash(),
                             System.currentTimeMillis(), "Block " + nextBlockIndex));
+
+            incrementBlockCount();
         } catch (NoSuchAlgorithmException e) {
             throw new RuntimeException(e);
         }
@@ -101,6 +125,8 @@ public class BlockchainService {
         }
 
         blockchain[nextBlockIndex] = block;
+        healthMetricService.incrementBlockCount();
+        System.out.println("incremented blockcount from blockchain service");
     }
 
     /**
@@ -148,12 +174,34 @@ public class BlockchainService {
      * @return The health metrics of the blockchain.
      */
     public Integer[] getHealthMetrics() {
-        Integer[] healthMetrics = new Integer[5];
 
-        for (int i = 0; i < healthMetrics.length; i++) {
-            healthMetrics[i] = (int) (Math.random() * 100 + 1);
-        }
+        // mock health metrics
+        contaminationSpectrum[0] = (int) (Math.random() * 100 + 1);
+        contaminationSpectrum[1] = (int) (Math.random() * 100 + 1);
+        contaminationSpectrum[2] = (int) (Math.random() * 100 + 1);
+        contaminationSpectrum[3] = (int) (Math.random() * 100 + 1);
+        contaminationSpectrum[4] = (int) (Math.random() * 100 + 1);
 
-        return healthMetrics;
+        Integer[] healthMetrics = Utils.getCopntaminationSpectrum(contaminationSpectrum);
+        Arrays.fill(contaminationSpectrum, 0);
+
+         return healthMetrics;
+    }
+
+    /**
+     * The incrementBlockCount method is responsible for incrementing the block count.
+     */
+    private void incrementBlockCount() {
+        contaminationSpectrum[0]++;
+    }
+
+    /**
+     * The getBlockCount method is responsible for returning the block count.
+     * @return The block count.
+     */
+    public int getBlockCount() {
+        int blocks = healthMetricService.getBlockCount();
+        healthMetricService.resetBlockCount();
+        return blocks;
     }
 }