+++ src/main/java/com/hunteryavitz/blockchainapi/utils/HealthMetricUtils.java
@@ -1,0 +1,5 @@
+package com.hunteryavitz.blockchainapi.utils;
+
+public class HealthMetricUtils {
+
+}